+++ src/main/java/kitchenpos/domain/order/OrderTableRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.domain.ordertable;
+package kitchenpos.domain.order;
 
-import kitchenpos.domain.ordertable.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
