+++ src/main/java/kitchenpos/domain/order/TableGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.domain.tablegroup;
+package kitchenpos.domain.order;
 
-import kitchenpos.domain.tablegroup.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
