+++ src/main/java/kitchenpos/domain/order/TableGroupValidator.java
@@ -1,8 +1,8 @@
-package kitchenpos.domain.tablegroup;
+package kitchenpos.domain.order;
 
 import kitchenpos.domain.order.OrderRepository;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.ordertable.OrderTable;
+import kitchenpos.domain.order.OrderTable;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
