+++ src/main/java/kitchenpos/ui/response/TableResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.response;
 
-import kitchenpos.domain.ordertable.OrderTable;
+import kitchenpos.domain.order.OrderTable;
 
 import java.util.List;
 import java.util.stream.Collectors;
